requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# UI
gradio>=4.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available
sentence-transformers>=2.2.0

# Optional GPU acceleration (comment out if not using)
# vllm>=0.3.0  # Requires CUDA, metal support for optimization

# Data processing
pandas>=1.5.0
PyYAML>=6.0
datasets>=2.14.0  # For HuggingFace dataset loading

# Logging and monitoring
python-dotenv>=1.0.0